More debug output Fix bug with dnotify backend and a subscription that is
* server/gam_poll.c: (gam_poll_scan_callback),
(gam_poll_scan_all_callback), (gam_poll_init_full),
(gam_poll_consume_subscriptions): More debug output
* server/gam_server.c: (gam_init_subscriptions),
(gam_add_subscription), (gam_remove_subscription): Fix bug with dnotify
backend and a subscription that is on the exclude list. What was happening
was that we were calling gam_poll_add_subscription in that case, but when
the dnotify backend is enabled the poll backend is in a different mode. So
these subscriptions was never consumed. I special cased for when inotify
is not running to call the backend_add_subscription even when the item is
on the exclude list. The dnotify backend has code to deal with excluded
subscriptions.